+++ packages/postcss-container-query/src/saveJSON.js
@@ -1,5 +1,27 @@
-import { writeFileSync } from "fs";
+import { readFile, writeFile } from "fs";
 
-export default function saveJSON(cssFile, json) {
-  writeFileSync(`${cssFile}.json`, JSON.stringify(json));
-}
+const saveJSON = (cssFile, json) =>
+  new Promise(resolve => {
+    const jsonData = JSON.stringify(json);
+    const jsonFilePath = `${cssFile}.json`;
+
+    const writeJson = () =>
+      writeFile(jsonFilePath, jsonData, e => {
+        if (e) {
+          console.error(`Failed to save container query json file: ${e}`);
+        }
+
+        resolve();
+      });
+
+    readFile(jsonFilePath, "utf8", (e, contents) => {
+      // Write file if it doesn't exist yet, or if the contents changed
+      if (e || contents !== jsonData) {
+        writeJson();
+      } else {
+        resolve();
+      }
+    });
+  });
+
+export default saveJSON;